Crate ecstasy

A simple ECS library.

Example

#[derive(Component, Debug, Deserialize, PartialEq, Serialize)]
struct Counter(usize);

#[system(simple)]
fn AssertNameHas3Bytes(entity: Entity, _dt: f32, name: &Name) {
    assert_eq!(name.0.len(), 3, "{:?}'s name ({}) should have been 3 bytes", entity, name);
}

#[system_mut(simple)]
fn IncrCounter(entity: Entity, _dt: f32, counter: &mut Counter) {
    counter.0 += 1;
}

struct SumDebugCounters<'a>(&'a AtomicUsize);
impl<'a> System for SumDebugCounters<'a> {
    fn run(&mut self, cs: &ComponentStore, _dt: f32) {
        cs.iter_entities().for_each(|entity| {
            if let (Some(counter), Some(_)) =
                    (cs.get_component::<Counter>(entity), cs.get_component::<DebugFlag>(entity)) {
                self.0.fetch_add(counter.0, Ordering::SeqCst);
            }
        })
    }
}

let mut n = AtomicUsize::new(0);
let mut engine = Engine::new()
    .add_mut_pass(IncrCounter)
    .build_par_pass()
        .add(AssertNameHas3Bytes)
        .add(SumDebugCounters(&n))
    .finish();

let foo = engine.store.new_entity();
let bar = engine.store.new_entity();
let baz = engine.store.new_entity();

engine.store.set_component(foo, Name("foo".to_string()));
engine.store.set_component(bar, Name("bar".to_string()));
engine.store.set_component(baz, Name("baz".to_string()));

engine.store.set_component(foo, Counter(0));
engine.store.set_component(bar, Counter(0));

engine.store.set_component(foo, DebugFlag);
engine.store.set_component(baz, DebugFlag);

while n.load(Ordering::SeqCst) < 25 {
    engine.run_once();
}

assert_eq!(engine.store.get_component::<Name>(foo).map(|Name(s)| &s as &str), Some("foo"));
assert_eq!(engine.store.get_component::<Name>(bar).map(|Name(s)| &s as &str), Some("bar"));
assert_eq!(engine.store.get_component::<Name>(baz).map(|Name(s)| &s as &str), Some("baz"));
assert_eq!(engine.store.get_component::<Counter>(foo), Some(&Counter(7)));
assert_eq!(engine.store.get_component::<Counter>(bar), Some(&Counter(7)));
assert_eq!(engine.store.get_component::<Counter>(baz), None);
assert_eq!(engine.store.get_component::<DebugFlag>(foo), Some(&DebugFlag));
assert_eq!(engine.store.get_component::<DebugFlag>(bar), None);
assert_eq!(engine.store.get_component::<DebugFlag>(baz), Some(&DebugFlag));
assert_eq!(n.load(Ordering::SeqCst), 28);

Re-exports

pub use ecstasy_proc_macros::system;

pub use ecstasy_proc_macros::system_mut;

pub use ecstasy_proc_macros::Component;

Modules

components
Some common components.

Structs

ComponentStore
A container for components.

Engine
Wraps a ComponentStore and several systems.

EnginePassBuilder
A builder for a new pass of Systems being added.

Entity
An entity.

Traits

Component
Components are data which can be attached to entities via a ComponentStore.

System
A system that does not modify the ComponentStore. These systems can be run in parallel with each other, but should generally not use parallelism internally.

SystemMut
A system that modifies the ComponentStore.