Struct ecstasy::ComponentStore

pub struct ComponentStore { /* fields omitted */ }

A container for components.

Methods

impl ComponentStore

pub fn new() -> ComponentStore

Creates a new, empty ComponentStore.

pub fn iter_entities(&self) -> impl Clone + Iterator<Item = Entity>

Returns an iterator over all entities.

pub fn new_entity(&mut self) -> Entity

Creates a new entity.

pub fn get_component<T: Component>(&self, entity: Entity) -> Option<&T>

Gets a component for a given entity.

pub fn get_mut_component<T: Component>(
&mut self,
entity: Entity
) -> &mut Option<T>

Gets a component for a given entity.

pub fn remove_component<T: Component>(&mut self, entity: Entity)

Removes a component from a given entity.

pub fn set_component<T: Component>(&mut self, entity: Entity, component: T)

Sets a component for a given entity.

pub fn take_component<T: Component>(&mut self, entity: Entity) -> Option<T>

Tries to remove a component from an entity.

pub unsafe fn unsafe_get_mut_component<T: Component>(
&self,
entity: Entity
) -> &mut Option<T>

Gets a component for a given entity. This is unsafe since it makes it possible to have two mutable references to the same component if called twice with the same T.

Safety

The references returned by calling this function with the same T must not exist at once.